\name{targets}
\alias{targets}
\docType{data}
\title{Targets file for gcspikelite data.}
\description{
  List of filenames and experimental group
}
\usage{data(targets)}
\format{
  Data frame with columns:
  \describe{
    \item{\code{Filename}}{a character vector}
    \item{\code{Group}}{a character vector}
  }
}
\keyword{datasets}
